+++ config/shell.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'adapters' => [
+        'local' => [
+            'class'     => \ConductorCore\Shell\Adapter\LocalShellAdapter::class,
+        ],
+    ],
+];