+++ src/Shell/ShellAdapterManagerFactory.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace ConductorCore\Shell;
+
+use ConductorCore\Exception;
+use Interop\Container\ContainerInterface;
+use Interop\Container\Exception\ContainerException;
+use Zend\ServiceManager\Exception\ServiceNotCreatedException;
+use Zend\ServiceManager\Exception\ServiceNotFoundException;
+use Zend\ServiceManager\Factory\FactoryInterface;
+use Zend\ServiceManager\ServiceManager;
+
+class ShellAdapterManagerFactory implements FactoryInterface
+{
+    /**
+     * Create an object
+     *
+     * @param  ContainerInterface $container
+     * @param  string             $requestedName
+     * @param  null|array         $options
+     *
+     * @return object
+     * @throws ServiceNotFoundException if unable to resolve the service.
+     * @throws ServiceNotCreatedException if an exception is raised when
+     *     creating a service.
+     * @throws ContainerException if any other error occurs
+     */
+    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    {
+        $shellAdapters = [];
+        $config = $container->get('config');
+        if (isset($config['shell']['adapters'])) {
+            foreach ($config['shell']['adapters'] as $name => $adapter) {
+                $class = $adapter['class'];
+                $arguments = !empty($adapter['arguments']) ? $adapter['arguments'] : [];
+                try {
+                    if ($arguments) {
+                        if ($container instanceof ServiceManager) {
+                            $shellAdapter = $container->build($class, $arguments);
+                        } else {
+                            throw new Exception\LogicException(
+                                'Adapter arguments not allowed if not using ' . ServiceManager::class . ' container.'
+                            );
+                        }
+                    } else {
+                        $shellAdapter = $container->get($class);
+                    }
+                } catch (ServiceNotCreatedException $e) {
+                    throw new Exception\RuntimeException("Error in shell/adapters/$name configuration");
+                }
+
+                $shellAdapters[$name] = $shellAdapter;
+            }
+        }
+        return new ShellAdapterManager($shellAdapters);
+    }
+}
+