+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file, in reverse 
 - Added Crypt commands and model
 - Added posix PHP extension suggestion
 - Added support for setting config on filesystems
+- Added shell adapters and commands
 
 # 0.1.0
 - Initial build copied over from devops tool